9.11.18 - Updates
Re writing loop as it was confusing and not working as expected. Sorting
volcanos into two arrays keeping the standard vol level array and making
a new array called inactive volcanos to loop though in order to create
icons for inactive ones if that is a desired feature. The subsucent
functiosn use the parents i as a index.