16.09.18 - Fixed Error
Fixed no map displaying when using lan cable and no address was being found. This caused the div called errorNotFound to hide the api. Since this seems to be a temp bug i have adjusted it so when exact location cant be found to just print no address found in address and not hide the map.